Fix sidebar overlapping with chapters on mobile For screens 580px wide or less: The sidebar (including zip download and feed buttons) will no longer  be partly hidden under the last chapter in the download-chapters table.
For screens between 581px and 660px wide:
That same sidebar will be below the chapter table, rather than above the
 book description and cover art.  This is consistent with how smaller
 screens have always worked (aside from the "partly-hidden" bug!).

The main page content was set to float left (which the sidebar is
supposed to do), but absolute positioning made that a moot point.  Now,
with both floating (and never overlapping), we can remove the
`margin-bottom` hack.

Tested on desktop versions of Firefox and Chromium, at various sizes.